Dedup spec account generator
